agents/requirements.txt
langgraph==0.2.35
pydantic==2.8.2
playwright==1.46.0
Pillow==10.4.0
numpy==2.0.1
